requirements.txt
streamlit==1.22.0
torch==2.1.0
torchvision==0.15.0
numpy==1.24.0
opencv-python==4.8.0.76
Pillow==9.5.0
